+++ convml_tt/trainer.py
@@ -6,9 +6,24 @@ import pytorch_lightning as pl
 from pytorch_lightning.loggers import WandbLogger
 from pathlib import Path
 import os
+import wandb
 
 from .system import TripletTrainerModel, TripletTrainerDataModule, HeadFineTuner
+from .interpretation.rectpred.sample import make_plot as rectpred_sample_plot
 from . import __version__
+
+
+def _make_rectpred_logger(image_path):
+    if image_path is None:
+        return lambda *args, **kwargs: None
+
+    import wandb
+
+    def _logger(model, stage):
+        fig = rectpred_sample_plot(model=model, image_path=image_path)
+        wandb.log({f"{stage}_rectpred": fig})
+
+    return _logger
 
 
 def main(args=None):
@@ -23,6 +38,12 @@ def main(args=None):
         default=False,
         action="store_true",
         help="Log training to Weights & Biases",
+    )
+    parser.add_argument(
+        "--sample-rectpred-plot-image-path",
+        default=None,
+        type=Path,
+        help="Use this image to produce a rectpred example at the beginning and end of training",
     )
     parser.add_argument(
         "--gpus", type=int, help="Number of GPUs to use for training", default=0
@@ -85,7 +106,12 @@ def main(args=None):
             {"convml_tt__version": __version__}
         )
 
+    rectpred_logger = _make_rectpred_logger(args.sample_rectpred_plot_image_path)
+    rectpred_logger(model=model, stage="start")
+
     trainer.fit(model=model, datamodule=datamodule)
+
+    rectpred_logger(model=model, stage="end")
 
 
 if __name__ == "__main__":